// ui/src/api/generated/model/smtpEncryption.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.7.1
 */

export type SmtpEncryption = typeof SmtpEncryption[keyof typeof SmtpEncryption];


export const SmtpEncryption = {
  starttls: 'starttls',
  ssl: 'ssl',
  none: 'none',
} as const;