// ui/src/api/generated/model/smtpProvider.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.7.1
 */

export type SmtpProvider = typeof SmtpProvider[keyof typeof SmtpProvider];


export const SmtpProvider = {
  aws_ses: 'aws_ses',
  google_workspace: 'google_workspace',
  google_workspace_relay: 'google_workspace_relay',
  custom_smtp: 'custom_smtp',
} as const;