// ui/src/api/generated/model/smtpUpdateRequest.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.5.1
 */
import type { SmtpEncryption } from './smtpEncryption';
import type { SmtpProvider } from './smtpProvider';

export interface SmtpUpdateRequest {
  provider: SmtpProvider;
  /** @maxLength 512 */
  host?: string;
  /**
     * @minimum 1
     * @maximum 65535
     */
  port?: number;
  encryption?: SmtpEncryption;
  /** @maxLength 512 */
  username?: string;
  /**
     * @minLength 1
     * @maxLength 2048
     */
  password: string;
  from_email: string;
  /** @maxLength 256 */
  from_name?: string;
  /**
     * @minimum 1
     * @maximum 300
     */
  timeout_seconds?: number;
}